+++ packages/gatsby-theme-aio/src/components/MDXFilter/MDXBlocks.js
@@ -21,6 +21,16 @@ import { CodeBlock } from '../CodeBlock';
 import { Variant } from '../Variant';
 import { TitleBlock } from '../TitleBlock';
 import { TextBlock } from '../TextBlock';
+import { Divider } from '../Divider';
+import { Carousel } from "../Carousel";
+import { TabsBlock } from "../Tabs";
+import { WrapperComponent } from '../WrapperComponent';
+import { FormWrapperComponent } from '../FormWrapperComponent';
+import { CustomMenuBlock } from '../CustomMenuBlock';
+import { CustomIframeBlock } from '../CustomIframe';
+import { CustomCard } from '../CustomCard';
+import { MenuWrapperComponent } from '../MenuWrapperComponent';
+import { Accordion, AccordionItem } from '../Accordion';
 import { AnnouncementBlock } from '../AnnouncementBlock';
 import { SummaryBlock } from '../SummaryBlock';
 import { ProductCard } from '../ProductCard';
@@ -28,8 +38,8 @@ import { ResourceCard } from '../ResourceCard';
 import { Media } from '../Media';
 import { JsDocParameters } from '../JsDocParameters';
 import { ProductCardGrid } from '../ProductCardGrid';
-import { TabsBlock } from '../Tabs';
-import { Carousel } from '../Carousel';
+import { AnchorButtonGroup } from '../AnchorButtonGroup';
+import { DCSummaryBlock } from '../DCSummaryBlock';
 
 export const MDXBlocks = {
   Hero,
@@ -40,6 +50,7 @@ export const MDXBlocks = {
   Variant,
   TitleBlock,
   TextBlock,
+  Divider,
   AnnouncementBlock,
   SummaryBlock,
   ProductCard,
@@ -47,6 +58,18 @@ export const MDXBlocks = {
   Media,
   JsDocParameters,
   ProductCardGrid,
-  TabsBlock,
-  Carousel
+  WrapperComponent,
+  FormWrapperComponent,
+  MenuWrapperComponent,
+  CustomMenuBlock,
+  CustomIframeBlock,
+  CustomCard,
+  Accordion,
+  AccordionItem,
+  TabsBlock: ({ theme, ...props }) => (
+    <TabsBlock theme={theme ?? globalTheme.code} {...props} />
+  ),
+  Carousel,
+  AnchorButtonGroup,
+  DCSummaryBlock
 };