+++ packages/gatsby-theme-aio/src/components/TitleBlock/index.js
@@ -14,10 +14,10 @@ import React, { cloneElement } from 'react';
 import { css } from '@emotion/react';
 import '@spectrum-css/typography';
 import PropTypes from 'prop-types';
-import { layoutColumns, TABLET_SCREEN_WIDTH } from '../../utils';
+import { layoutColumns, TABLET_SCREEN_WIDTH, DESKTOP_SCREEN_WIDTH } from '../../utils';
 import classNames from 'classnames';
 
-const TitleBlock = ({ className, heading, text, theme = 'lightest' }) => (
+const TitleBlock = ({ className, heading, text, theme = 'lightest',id }) => (
   <section
     className={classNames(className, `spectrum--${theme}`)}
     css={css`
@@ -64,7 +64,8 @@ const TitleBlock = ({ className, heading, text, theme = 'lightest' }) => (
 TitleBlock.propTypes = {
   heading: PropTypes.element,
   text: PropTypes.element,
-  theme: PropTypes.string
+  theme: PropTypes.string,
+  id: PropTypes.string,
 };
 
 export { TitleBlock };